requirements.txt
fastapi
uvicorn
langchain_groq
python-dotenv
pydantic
requests